Update contact details and package dependencies
The contact details in the README.md file have been updated to reflect a change in project ownership. References to the previous owner, salman0ansari, have been replaced with references to the new owner, jlmacedo7. In package.json file, the package version is updated to 3.0.9, and various dependencies have been updated to newer versions. The repository and author fields have also been updated.