Update README.md with new project leadership details
The README.md file has been edited to reflect a change in project leadership. Messages referring to the origination and archiving of the project by the previous owner have been replaced by new messages from the current support provider. The call for support to maintain the project has been included.